deletion logic!

+++ src/zipoc/cli.py
@@ -43,7 +43,19 @@ def repository_delete():
     if repo_path.exists():
         print("Zipoc repository hasn't been initialized!")
         return 0
-    os.remove(repo_path)
+    confirm = input('Are you sure? Deleting a repository is irreversible and your data will be lost! (y/n)')
+    if confirm == 'n':
+        print("Zipoc repository deletion cancelled!")
+        return 0
+    else:
+        try:
+            os.remove(repo_path)
+            print("Successfully deleted zipoc repository!")
+            return 0
+        except Exception as e:
+            print("Unknown error while trying to delete zipoc repository.")
+            return 0
+
     
     return 0
